refactor: Extract shared visualization primitives to lib/visualization
Move star-parser, saga-mermaid, and graph layout helpers from
features/cookbook/lib to lib/visualization for reuse across features.

- Move star-parser.ts and saga-mermaid.ts (with tests) to lib/visualization/
- Extract layoutWithELK, node dimension constants, and edge style constants
  from composition-graph.tsx into lib/visualization/graph-layout.ts
- Add DEFAULT_MAX_CHAIN_DEPTH constant in lib/visualization/constants.ts
- Create barrel export at lib/visualization/index.ts
- Replace old cookbook files with re-exports for backwards compatibility
- Update composition-graph.tsx to use shared graph-layout module
